requirement_prod.txt
annotated-types==0.6.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-ninja==1.1.0
django-ninja-extra==0.20.7
dnspython==2.6.1
email_validator==2.1.1
h11==0.14.0
idna==3.7
injector==0.21.0
mysqlclient==2.2.4
ninja-schema==0.13.6
pydantic==2.7.1
pydantic_core==2.18.2
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.30.1
requests~=2.32.3